test(*): fix jest config

+++ tests/filesystem/filesystem.test.ts
@@ -1,6 +1,6 @@
-import { FileSystemBucketConfig, FilesystemProvider, FileSystemProviderConfig } from '../../packages/file-storage/filesystem/src';
+import { FilesystemProvider, FileSystemProviderConfig } from '../../packages/file-storage/filesystem/src';
 import path from 'path';
-import { Bucket, File, FileStorage, Streams } from '../../packages/file-storage/core/src';
+import { Bucket, StorageFile, FileStorage } from '../../packages/file-storage/core/src';
 import https from 'https';
 import fs from 'fs';
 
@@ -8,7 +8,14 @@ describe('Filesystem tests', () => {
     const rootPath = path.resolve('./private/filesystem/');
 
     beforeAll(() => {
-        fs.rmSync(rootPath, { recursive: true });
+        if (fs.existsSync(rootPath)) {
+            fs.rmSync(rootPath, { recursive: true });
+        }
+    })
+    afterAll(() => {
+        if (fs.existsSync(rootPath)) {
+            fs.rmSync(rootPath, { recursive: true });
+        }
     })
 
     FileStorage.registerProviderType('fs', FilesystemProvider);
@@ -55,13 +62,27 @@ describe('Filesystem tests', () => {
         })).result;
         expect(bucket04).toBeInstanceOf(Bucket);
 
+        expect((await bucket04.removeEmptyDirectories()).result).toBeTruthy();
+
+        // check can read/ can write
+        expect(bucket01.canRead()).toBeTruthy();
+        expect(bucket01.canWrite()).toBeTruthy();
+
 
         // put files
         const file01 = (await bucket01.putFile('file01.txt', 'Test 01')).result;
         expect(file01).toBe('provider01://bucket01/file01.txt');
-        expect((await bucket01.fileExists(file01, true)).result).toBeInstanceOf(File);
+        expect((await bucket01.getFile('file01.txt')).result.getStorageUri()).toBe(file01);
+        expect((await bucket01.getFile('file01.txt')).result.getPublicUrl()).toBeTruthy();
+        expect((await bucket01.getFile('file01.txt')).result.getSignedUrl()).toBeTruthy();
+
+        // check native path
+        expect(bucket01.getNativePath(file01)).toBe(path.join(bucket01.provider.config.root, bucket01.config.root, 'file01.txt'));
+
+
+        expect((await bucket01.fileExists(file01, true)).result).toBeInstanceOf(StorageFile);
         const file02 = (await bucket02.putFile('file02.txt', 'Test 02', { returning: true })).result;
-        expect(file02).toBeInstanceOf(File);
+        expect(file02).toBeInstanceOf(StorageFile);
         expect((await bucket02.fileExists(file02)).result).toBe(true);
 
         const p = new Promise(resolve => {
@@ -71,18 +92,19 @@ describe('Filesystem tests', () => {
             });
         });
 
-        expect(await p).toBeInstanceOf(File);
+        expect(await p).toBeInstanceOf(StorageFile);
 
         // copy between buckets
         const file03 = (await bucket01.copyFile('/file01.txt', 'provider01://bucket02/file_copy.txt', { returning: true })).result;
-        expect(file03).toBeInstanceOf(File);
+        expect(file03).toBeInstanceOf(StorageFile);
 
         const file04 = (await bucket01.copyFile('/file01.txt', 'provider01://bucket02/', { returning: true })).result;
-        expect(file04).toBeInstanceOf(File);
+        expect(file04).toBeInstanceOf(StorageFile);
+        expect(file04.getNativePath()).toBe(path.join(file04.bucket.provider.config.root, file04.bucket.config.root, file04.getAbsolutePath()));
 
         // move between buckets
         const file05 = (await bucket02.moveFile(file02, 'provider01://bucket01/subdir 01/file moved.txt', { returning: true })).result;
-        expect(file05).toBeInstanceOf(File);
+        expect(file05).toBeInstanceOf(StorageFile);
         expect((await bucket02.fileExists(file02)).result).toBeFalsy();
 
         // list files
@@ -99,6 +121,7 @@ describe('Filesystem tests', () => {
         bucket01Filelist = (await bucket01.listFiles('/', { recursive: true, pattern: '/subdir*/**/*.txt', returning: true })).result;
         expect(bucket01Filelist.entries).toHaveLength(1);
         expect((await bucket01Filelist.entries[0].getContents()).result.toString()).toBe('Test 02');
+        expect((await bucket01Filelist.entries[0].getStream()).result).toBeTruthy();
 
         const bucket02Filelist = (await bucket02.listFiles()).result;
         expect(bucket02Filelist.entries).toHaveLength(2);
@@ -124,6 +147,11 @@ describe('Filesystem tests', () => {
 
         expect((await bucket02.deleteFiles('/', '**/subdir-01/*.txt')).result).toBe(true);
         expect((await bucket02.listFiles('/', { recursive: true })).result.entries).toHaveLength(3);
+
+        // destory a bucket
+        expect((await bucket04.destroy()).result).toBeTruthy();
+        expect((await bucket02.remove()).result).toBeTruthy();
+
 
     });
 });
